requirements.txt
python-dotenv
scrapy
parsel
requests
twisted
pymysql
